Privacy: Disable tracking of recent files by default
A memory leak was found in the refreshing of recent files in the menu applet.

Although the impact was of the leak was greatly reduced, it still is present.

Until it is completely fixed, we should favor speed and performance
over the ability to remember old files.

So this option is now disabled by default.