Meson deprecations (#128)
* meson: require the minimum meson version for features we use

pkgconfig custom variables are not available in 0.37

* meson: fix deprecation warnings

build_always is replaced in spirit by build_always_stale, and had the
side effect of activating build_by_default which was only available in
meson >= 0.38. Since we already require a higher meson version, and the
behavior we actually want is the latter, we should just use it directly.