Add gnome-systemd
This was added by GNOME in gnome-desktop to make it easy to create
scopes and run processes with predefined defaults.

The idea was to be able to lower processes timeout when killed or to
implement OOM features.

More info at:

https://gitlab.gnome.org/GNOME/gnome-desktop/-/merge_requests/58
https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/863
https://gitlab.gnome.org/GNOME/gnome-session/-/merge_requests/32
https://gitlab.gnome.org/GNOME/gnome-settings-daemon/-/merge_requests/147

https://github.com/systemd/systemd/issues/14202


This also results in moving processes under the Cinnamon process (which
then acts as parent process).

Unless/until this fixes an actual identified issue for us
or provides significant advantages we're not using it in Cinnamon.

The decision was nonetheless to leave it here in cinnamon-desktop to
reduce the difference with gnome-desktop for maintenance purpose and
to have it in if we ever want to use it in the future.