Add gnome-idle-monitor
Backported from gnome-session.

This is used in cinnamon-session, as a replacement for CSM's obsolete
version (cs-idle-monitor), which didn't support Wayland.

This one supports both Xorg and Wayland.

Even though it's only used in CSM, we have it here in cdesktop to
reflect the fact that it was added to gdesktop (easier maintenance and
backports).