Don't show keyboard layout flags by default
In the past we only supported layouts, now we also support
input modes.

Say you're using ibus/mozc for Japanese. We don't just show JP,
we show a hiragana or katakana or A symbol depending on the input mode.

Flags can't really show that, so that leads to a mixture of flags
and symbols for some users. The best OOTB experience is thus to not
use flags here.