gnome-xkb-info: Don't rely on gsettings to return the full set of xkb layouts and options.
Being able to only show the most common layouts in the Add Layout
dialog is useful, but Cinnamon's KeyboardManager and cinnamon-
settings should always be working with the full set.

Get rid of checking a setting and just add another constructor
that always returns the full list. The Add Layout dialog can still
use the setting to determine how many layouts it presents to the
user.